chore: divide fastapi in mode files
